Nakamoo/voice op

+++ client/managers/youtube.py
@@ -3,12 +3,20 @@ import subprocess
 from subprocess import Popen
 import requests
 from bs4 import BeautifulSoup
+import logging
 
+import sys, os
+pardir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(pardir)
+import utils.tts as tts
+from utils.speechrecognition import confirm
 import traceback
 
 class Manager:
     def __init__(self, user, server_ip, actions):
         self.now_playing = None
+        self.user = user
+        self.ip = server_ip
 
     def start(self):
         pass
@@ -16,14 +24,36 @@ class Manager:
     def execute(self, acts):
         for act in acts:
             if act["platform"] == "play_youtube":
+                if "confirmation" in act:
+                    logging.debug("confirm: " + act['confirmation'])
+                    ans = confirm(act['confirmation'])
+                    logging.debug("answer: " + str(ans))
+                    data_confirm = {'platform': act['platform'], 'data': act['data'], 'user_name': self.user,
+                            'confirmation': act['confirmation'], 'answer': ans}
+
+                    if ans is None:
+                        tts.say("上手く聞こえませんでした")
+                        return
+                    elif not ans:
+                        tts.say("わかりました，再生をキャンセルします")
+                        return
+
+                    r = requests.post("%s/data/confirmation" % self.ip, data=data_confirm, verify=False, timeout=1)
+                    logging.debug(r)
+
                 try:
                     Popen('pkill -9 mpv', shell=True)
                     time.sleep(0.3)
+                    tts.say(act['data'] + "を検索します")
                     try:
                         youtube_result = Youtube(act['data'], result=1)
-                        Popen("mpv '" + youtube_result.url[0] + "' --loop --no-video > /dev/null 2>&1", shell=True)
+                        logging.debug('play: '+ youtube_result.url[1])
+                        Popen("mpv '" + youtube_result.url[1] + "' --loop --no-video", shell=True)
+                        # Popen("mpv '" + youtube_result.url[0] + "' --loop --no-video > /dev/null 2>&1", shell=True)
                     except:
-                        Popen("mpv https://www.youtube.com/watch?v=HKKe7p44PDY --loop --no-video > /dev/null 2>&1", shell=True)
+                        traceback.print_exc()
+                        tts.say(act['data'] + "は見つかりませんでした")
+                        # Popen("mpv https://www.youtube.com/watch?v=HKKe7p44PDY --loop --no-video > /dev/null 2>&1", shell=True)
                     self.now_playing = act['data']
                 except:
                     traceback.print_exc()
@@ -44,33 +74,10 @@ class Youtube():
 
         self.data = [h3 for h3 in h3s]
         self.url = ["https://www.youtube.com" + h3.a.get('href') for h3 in h3s]
-        self.title = [h3.a.get("title") for h3 in h3s]
-        self.id = [h3.a.get("href").split("=")[-1] for h3 in h3s]
-        self.embed = ["https://www.youtube.com/embed/" + h3.a.get("href").split("=")[-1] for h3 in h3s]
-        self.time = [h3.span.text.replace(" - 長さ: ","").replace("。","") for h3 in h3s]
-        self.info = [h3.text for h3 in h3s] # >>タイトル　- 長さ：00:00。
-
-    def select(self):
-        values = {"url":self.url,"title":self.title,"id":self.id,"embed":self.embed,"time":self.time}
-        info = self.info
-        for i in range(len(info)):
-            print("%s:%s" % (i,info[i]))
-        while True:
-            try:
-                num = int(input("番号:"))
-                break
-            except:
-                print("番号を正しく入力してください。")
-        results = {
-            "url":values["url"][num],
-            "title":values["title"][num],
-            "id":values["id"][num],
-            "embed":values["embed"][num],
-            "time":values["time"][num],
-            }
-        return results
 
 if __name__ == "__main__":
-    youtubeplayer = Manager('sample', '1.0.0.0', None)
-    youtubeplayer.execute([{'platform': 'play_youtube', 'data': 'n'}])
+    youtubeplayer = Manager('koki', 'https://homeai.ml:5007', None)
+    youtubeplayer.execute([{'platform': 'play_youtube', 'data': 'smap', 'confirmation': 'smapを再生しますか?'}])
     # youtubeplayer.execute([{'platform': 'stop_youtube', 'data': ''}])
+    while True:
+        pass
